Made multiple changes to make the AppImage hopefully work on more systems

+++ Src/Program.cs
@@ -40,13 +40,18 @@ internal static class Program
     {
         var path = GetExecutablePath();
         var currentProcess = Process.GetCurrentProcess();
-        foreach (var process in Process.GetProcesses().Where(p => !p.Equals(currentProcess) && p.MainModule?.FileName == path))
+
+        // Only kill orphaned processes in foreground mode (not in background --bg mode)
+        // This prevents double AppImageLauncher prompts
+        if (!(args.Length > 0 && args[0] == "--bg"))
         {
-            Console.WriteLine("Killing previous orphaned process");
-            try
+            foreach (var process in Process.GetProcesses().Where(p => !p.Equals(currentProcess) && p.MainModule?.FileName == path))
             {
-                process.Kill(true);
-            } catch (Exception) { /* nom */ }
+                try
+                {
+                    process.Kill(true);
+                } catch (Exception) { /* nom */ }
+            }
         }
 
         if (args.Length > 0 && args[0] == "--bg")
@@ -69,25 +74,7 @@ internal static class Program
 
     // Avalonia configuration, don't remove; also used by visual designer.
     private static AppBuilder BuildAvaloniaApp()
-        => AppBuilder.Configure<App>()
-            .UsePlatformDetect()
-            .With(new Win32PlatformOptions
-            {
-                CompositionMode = [Win32CompositionMode.LowLatencyDxgiSwapChain, Win32CompositionMode.RedirectionSurface],
-                RenderingMode = [Win32RenderingMode.AngleEgl, Win32RenderingMode.Software],
-                ShouldRenderOnUIThread = false,
-            })
-            .With(new X11PlatformOptions
-            {
-                EnableIme = false,
-                EnableMultiTouch = false,
-                EnableSessionManagement = false,
-                RenderingMode = [X11RenderingMode.Glx, X11RenderingMode.Software],
-                ShouldRenderOnUIThread = true,
-                UseRetainedFramebuffer = false,
-            })
-            .With(new SkiaOptions
-            {
-                UseOpacitySaveLayer = true,
-            });
+    {
+        return AppBuilder.Configure<App>().UsePlatformDetect();
+    }
 }
